backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai>=1.0.0
sqlmodel